refactor : 데이터 Soft Delete 방식으로 수정

+++ src/main/java/com/example/board/domain/post/dto/PostListResponse.java
@@ -15,14 +15,14 @@ public class PostListResponse {
 
     private LocalDateTime createdAt;
 
-    private Integer commentSize;
+    private Long commentSize;
 
     public PostListResponse() {
     }
 
     @QueryProjection
     public PostListResponse(Long id, String title, String content, LocalDateTime createdAt,
-        Integer commentSize) {
+        Long commentSize) {
         this.id = id;
         this.title = title;
         this.content = content;
